row sparse allreduce

+++ scripts/language_model/large_word_language_model_hvd_v1.py
@@ -215,7 +215,7 @@ def train():
     model.initialize(mx.init.Xavier(factor_type='out'), ctx=ctx)
     trainer_params = {'learning_rate': args.lr, 'wd': 0, 'eps': args.eps}
     # trainer = gluon.Trainer(model.collect_params(), args.optimizer, trainer_params)
-    trainer = DistributedRspTrainer(model.collect_params(), args.optimizer, trainer_params)
+    trainer = DistributedRspTrainer(model.collect_params(), args.optimizer, trainer_params, sdtype='float16')
 
     if args.from_epoch:
         from_epoch = args.from_epoch
